reorganise tests relating to np interop
I have gathered all the numpy related tests into a single test file,
test_npsupport.py, which has a couple of different TestCases internally,
reflecting the original organisation. test_npsupport will only be run if
importing numpy succeeds.

I also removed a lot of @skip decorators relating to numpy dependence, all
tests currently pass (or are skipped) with or without numpy being installed
(on my system at least, Python 3.9.13 and numpy 1.23.1).
